Created Methods to update Box Scopes
All scopes will import the pc object from the worldstate module,
which will pass the current_room and items_worn to the scopes.

If the creation of the pc object changes location, the import will
need to be changed. Alternatively, we can just the pass self to the
scope updates through player methods when they take actions to
change each scope.

I've also set it currently to have any scope update call for a
full localscope update, which will call roomscope and selfscope. If
this becomes a huge performance issue, this can be changed.