Added two new text methods
Numstr takes a plain integer from 1 to 10 and returns the verbose
version of the number. Not super necessary but it does add a little
craftsmanship to the item display methods and anything else I need
it for.

Added a default textwrap method too using textwrap. May end up
expanding it later for menus and other things, especially the
dialogue screens.