Added functionality for items and containers
Added two args for Box __init__ to determine if box is closed and
locked. Also added a list description for all boxes to display
when getting, putting, or listing items. Added opening and closing
of boxes, and then added the below-mentioned list items to the box
description.

Added the list items method to greatly improve item and room
descriptions. The method will search itself for all items, and
then search all open boxes recursively and display those items
in their own strings.

Added search_scope for a box itself so that the search scope can
be called now on a scope or just a box.