Added item and container functionality
First, I reorganized the methods in the player class.

Added a general open/close/unlock method to be called in gameloop,
and then depending on what object was passed, the method will
determine whether to pass to a box or a door.

Finished up get_item for boxes, and then added a drop and place
item method to the player.