Added Functionality to Wear Items
Player is now able to grab a wearable item from the world and put
it on. These include backpacks. These items remain in the player
inventory attribute, but when listing inventory in the menu, the
equipped items should be ignored (for a future addition).

Adding the backpacks made me realize that when items were in the
player's inventory they just kind of dissappeared unless they were
in a worn container. This is because I forgot to add the player's
inventory to selfscope. Because of how the boxscopes work, if I
wear an item, it falls off the searchscope.

This means you couldn't look at or otherwise interact with equipped
items, which as of now wouldn't be a major issue if it weren't for
the fact that to put items in a backpack, it needs to show up as an
item in the selfscope.

That was easily fixed by keeping worn items in the player's
"inventory". I'll find some other way to limit "held items".