Blocked overworld scenescope methods
Since a maproom doesn't have an inventory, players shouldn't be
dropping or otherwise trying to interact with that map space. So
When entering an overworld maproom the scenescope will be cleared,
and get/place methods will not search it for items and containers.
However, dropping items will work but it will basically remove them
from the game.

There was also a bug where trying to get an item from the room while
an item with the same id from your inventory was throwing a
"multiple item with same name" game message, so I had to mess with
the localscope update method to make sure that the scenescope didn't
grab the selfscope.